giscanner: Handle pid_t, etc. as their own top-level types
Previously we assumed that these are int-sized, but that isn't necessarily
true on every Unix variant (and in fact dev_t is 64-bit on x86 Linux,
so was always wrong here). Give them the same treatment as time_t.

These types are not included in INTROSPECTABLE_BASIC and therefore not
included in `everything.[ch]`, because the code that generates those
files is not set up to be able to deal with types that only exist on
certain platforms. They *are* included in `gimarshallingtests.[ch]`,
but only within `#ifdef G_OS_UNIX`.

Resolves: https://gitlab.gnome.org/GNOME/gobject-introspection/-/issues/506
Signed-off-by: Simon McVittie <smcv@debian.org>